Separate \global
Move \newcommand back to macros for now
